requirements.txt

dbt-core
dbt-postgres
dagster-postgres
pandas
requests
beautifulsoup4
python-dotenv
mistralai
crawl4ai
langgraph
langchain-ollama
langchain-core
psycopg2-binary
langchainhub
streamlit
plotly
